manage codeStyles PACKAGES_IMPORT_LAYOUT IDE bug
See:
https://stackoverflow.com/questions/63656836/android-studio-codestyles-project-xml-configuration,
https://youtrack.jetbrains.com/issue/KT-39937

Attempting to revert the change results in the IDE regenerating it
again. So, if we have to leave this change out, we have to manually
remember to exclude it from every commit we make which isn't ideal.
Adding a codeStyle file to .gitignore isn't ideal either.

For now, we can just put these bugged changes as a separate commit so
that it can be reverted later when the bug gets fixed or some other
solution is found.