Patient: test null and empty Strings explicitly
We run into an issue: For a property P of a type T where the validity
of values for P depends on the values of other properties in an instance
of T, we can't reliably test the validity of a value for P without
creating an instance for T.

For example, for a Patient, a null age is valid iff the patient has a
date of birth. So, if we want to determine if a null age is valid, we
need an instance of a Patient. However, if we're filling out a form and
we want to do real-time validation as the user enters, this would mean
creating new Patient instances every time the user enters something.
This is inefficient.