implement age input and move error setting to XML
The date picker is now integrated into the text input, so it's hopefully
clear what the user will be choosing

Other changes:
* Use Material TextInputLayout for errors
* Make more room for id, age
* Move calendar icon to start, because the error icon overrides the end
  icon, which isn't ideal.
* Don't data bind age input, because the Fragments seem to be broken
  with Data Binding right now and it seems too complicated to handle
  at the moment.
* move error messages into XML via Data Binding
Use a map of error messages. This lets us clear error messages for
other attributes if we need to.
* Clean up age input state
* use proper theme for Drawable
The theme needs to be derived from the EditText's theme, so we have to
use the Context from there. Using it from there also seems to guarantee
that the Context is non-null
* add date picker constraints
* refactor age setup
* correctly report date errors
* display age for date of birth
* stub out broken BindingAdapters
* store calculated age if date of birth is input
* start Data Binding age input again
The UI for age input has been changed to show the age if the user inputs
a date of birth. This means that the age field gets overwritten.

We should note that the approximate age is not robust to time, since
it's just stored as an absolute number. It really should be stored as a
timestamp. However, the front end sets an approximate date of birth for
the age. More thought needs to be put into decide if we want to emulate
that behavior or if we need to change it (since it can be misleading to
have an approximate date of birth but no indication that it is
approximate)

* explain why we use setText despite Data Binding
* explain onAgeStateChange, and refactor date of birth Boolean
* update helper text to reflect date of birth usage
* show invalid date of birth
We need to set the errors after onAgeStateChange, as that can change
the helper text which interferes with the error text.
* fix age error message logic
* migrate age input state to a LiveData instance

* solve critical section with TextInputLayout errors
If a thread sets the error text for a TextInputLayout to a non-null
String, and then another thread sets the helper text for a
TextInputLayout to a non-null string, then the helper text will take
priority. To solve this, we use a Mutex and launch editing of those
fields in a coroutine. The Mutex.lock function is a suspend function, so
it won't block any threads.