fix BindingAdapters to work with Fragments
Apparently, we need to use a DataBindingComponent for Fragments, and
then the binding needs a LifeCycle owner set. The Fragment argument
seems optional, but we can keep it for now in case we need to use it in
the future.

See https://github.com/android/architecture-components-samples/tree/
master/GithubBrowserSample/app/src/main/java/com/android/example/github/
binding.