don't do Data Binding for AutoCompleteTextView
Data Binding on AutoCompleteTextView is broken when trying to set a
current menu item, because using it properly requires using setText
with the filter parameter set to false, but the default BindingAdapter
for AutoCompleteTextView doesn't do that when it calls setText. We could
copy the BindingAdapter for TextView changes to add that in, but that
would be a lot of copied code.