implement gestational age input
Fix GestationalAge unit LiveData not updating by observing it in the
PatientInfoFragment.

* Implement null bails again in the MediatorLiveData and shorten errors
* Convert gestational age in real time if user switches the units. This
differs from the frontend behavior.
* Disable + clear gestational age when not pregnant
* Show gestational age input mandatory when pregnant
* Display error if user keeps gender as null
* Make various changes to errors and BindingAdapters
* Shorten missing gestational age error, since it fails to fit in the
small space allocated for it.
* Don't duplicate stars indicating mandatory entry
* Ensure Doubles are also truncated in initial state
* Remove redundant "*Required" helper texts
* Micro-optimize error message setting
* Explain "bind" XML namespace for BindingAdapters
* zero out gestational age input when switching unit
We previously had some attempt to convert, but all these time classes
(Months, Weeks) seem to have floating-point precision errors. Zeroing
the input is also what the frontend does, so it's more in line with that
now. Ideally we could use something like BigDecimal to ensure precision.
* explain certain usages of MediatorLiveData sources
* refactor gestation age MediatorLiveData setup
* fix jumpy behavior when first entering month
When trying to input months for the first time,
`patientGestationalAgeInput` still has `patientGestationalAge` as a
LiveData source. This results in the DecimalFormat code running. We can
fix this by only showing the necessary decimal places.

* clarify some strings
The label for patient ID was missing the star, and we should clarify
that the icon for date of birth input can be clickable.

* Set default gestational age units more obviously
* Prevent animation popping due to empty value
* Limit gestational age input lengths