properly unmarshal drug and medical history
The server returns and accepts Strings for drug and medical history.
The frontend currently makes these text fields as paragraph fields, not
list entries. Treating the response as an Array field always results in
null being returned, hence, for now, we treat them as Strings and put
them into single-element Lists.

We need to do one of the following:
* Get the frontend and backend to treat these as Arrays. That implies
  work on the frontend to get them to use a list interface for inputting
  medical and drug history, and maybe server-side database migrations
* Migrate the Android database to store medical and drug history as
  Strings
* Interpret the server's string response as a List of Strings somehow
  (unreliable, because the frontend allows arbitrary formatting)