add SymptomsState to keep track of symptoms
The server receives and sends symptoms in English. If we want the
Android app to support i18n, this isn't really going to work.

So, for now, we leave the symptoms in the Reading class as a List of
Strings in English, and we implement SymptomsState class to track the
symptoms state independent of the English strings that are used. The
symptoms as defined by the symptoms string array in strings.xml
correspond index-wise to the BooleanArray inside a SymptomsState object,
where true means the symptom was present in the reading and false means
not present.

A SymptomsState object can accept an array of all default symptoms for
any language and give a list of the selected symptoms that's in the
given language. SymptomsState also maintains the property that in the
BooleanArray, the first value (corresponding to the "No symptoms"
checkbox) is true if and only if all the other values in the
BooleanArray and otherSymptoms is blank (not present). In this way, it
extracts all the symptom checkbox-related logic.

It's possible to construct a SymptomsState object from a list of English
symptoms along with the hardcoded String array of symptoms in English.
There's a constructor that parses that list of symptoms in English and
creates the appropriate SymptomsState object. From that, we can plug in
an array of symptoms in any language (with the array corresponding to
the original English array of course) to get a list of symptoms that are
selected in that language.