Reading flow: add patient-ID-already-exists dialog
This involves a database query and possibly a server check to see if the
patientId is already being used. To facilitate this, we let the
ViewModel control whether the EditText and other user input is disabled
so that the user doesn't not alter anything while we are checking the
patient ID. We also show a message in the bottom navigation bar of the
ReadingActivity when a reading ID check happens.

Things left to do: Use the actual patient, and download and add the
remote patient on the server.