support ReadingActivity with just patientId
The core initialization logic was copied over from the old code for
legacy reasons, but it's unclear why a readingId is required when a
patientId is sufficient for creating a new reading for a patient. It's
just unnecessary database accesses.

For making an Intent for adding a new reading to an existing patient, we
deprecate the maker that depends on a readingId and encourage the usage
of the new Intent maker that depends on a patientId.