use simpler logic for existing patient readings
There's no reason why a patient should need a reading in order to make a
reading. This would mean that if there was a patient from the server
that had no readings, it would be impossible to make new readings for
them in the PatientProfileActivity.

The reading IDs should just be UUIDs and shouldn't depend on previous
readings (unless it's for a rechecking vitals scenario).

This means we can finally get rid of the old Intent maker.