migrate to Hilt for dependency injection
Hilt is built on top of Dagger and makes it easier to work with Android
without needing to deal with boilerplate code.

For example, to get Hilt to inject into a standard Android component
(Activity, Fragment, Service, View, etc.), annotate the class with
`@AndroidEntryPoint` and it will inject the lateinit var fields for us.
If a Fragment requires injection, the hosting Activity will also need
the annotation, even if the Activity doesn't need injection.
(For Application, we annotate with @HiltAndroidApp`.)

In all other cases like plain-old Kotlin / Java objects, it's better to
get things injected via constructor injection. Otherwise, we can define
an entry point to get things injected.

See SyncStepperImplementation for an example of how entry points are
used for injection into a non-Android component. (It actually seems
better than before, since we can now maintain privacy of those variables
and mark them as val instead of var.)

References:
https://dagger.dev/hilt/migration-guide
https://codelabs.developers.google.com/codelabs/android-hilt/index.html
https://codelabs.developers.google.com/codelabs/android-dagger-to-hilt/

* Remove redundant `@Singleton` annotations from DAOs, because they're
  already coming from a singleton
* Use proper way of creating ViewModels for the
  HealthFacilitiesViewModel, and inject into it using ViewModelInject
* Move MyApp to the base package, since it makes more sense there and
  we don't have to asssociate it with dependency injection
* Remove Dagger setup from MyApp, because Hilt handles that for us
* Move everything related to dependency injection into the "di" package,
  which stands for dependency injection
* Remove Dagger dependencies in our gradle configuration, since Dagger
  is no longer used.