Reading flow: go back to start after process dies
This is a workaround. A proper solution would be to somehow persist all
the user input somewhere. Right now, when the system kills the app
process and the user starts the app up again, we move the user back to
the start destination, since all the fields have been cleared due to the
process death.

While we don't really expect this to happen too much since a user
creating a reading is expected to complete it through in one sitting, we
need to handle this case somehow.