Reading flow: address Next button critical section
It's possible after a system-initiated process death that when the
Android OS kills the app while we are in the VitalSignsFragment, the
Next button does not activate. This is since vitalSignsFragmentVerifyJob
launches and overrides the button state after code to change the next
button criteria to the Symptoms criteria runs.

We also now make use of the Mutex.withLock extension function.