Patient: implement workaround for patientAge issue
The API no longer works with patient age (a good thing, because it's
just based on an absolute number that doesn't change which is not how
time works). However, the Android app is out of date. Furthermore, the
server doesn't indicate whether a received date of birth is approximate
or not.

Therefore, we need some workaround until the isValidDob is available in
the API, and only then can we commit to changing the Patient schema.