Reading flow: adjust scrolling EditText behaviors
It seems like it's a losing battle to have a scrollable EditText inside
of a ConstraintLayout that is nested inside a ScrollView, because the
priority always seems to go to the ScrollView. Rather than dealing with
it, we just make the drug and medical history fields expand as much as
possible up to a certain reasonable limit.