delete Montserrat and add Open Sans Bold directly
Leaving the Open Sans Bold font / typeface to a font provider can result
in issues, since some Android phones may not have Google Mobile Services
and it will not show the correct typeface for those phones.

Montserrat is also not used anywhere and is not part of our style guide,
so we delete it.

Open Sans Bold is under Apache 2.0 and obtained from
https://fonts.google.com/specimen/Open+Sans
?query=open+sans&sidebar.open=true&selection.family=Open+Sans:wght@700