change version to 0.1.0-alpha
We consider the app and the entire CRADLE platform under the alpha stage
at the moment, so we remove the "Beta" term.

Prior to this commit, the Google Play Store version was 19, which wasn't
published to the repo. We have to increment the version code by 2 to get
over that.