PatientsActivity: use Paging 3 library for list
This significantly improves the performance of the patients list.

"PagingDataAdapter is a RecyclerView.Adapter base class which can
present the content of PagedLists in a RecyclerView. It requests new
pages as the user scrolls, and handles new PagedLists by computing list
differences on a background thread, and dispatching minimal, efficient
updates to the RecyclerView to ensure minimal UI thread work" [1].

References:
[1]: https://github.com/android/architecture-components-samples/blob
/a2151cf66483560f76b041ba95ce96e931c50caf/PagingSample/app/src/main/java
/paging/android/example/com/pagingsample/CheeseAdapter.kt#L26
[2]: https://developer.android.com/codelabs/android-paging