LoginManager: parse JSON from Http's InputStream
This improves the memory usage and prevents OutOfMemory exceptions
caused by storing a large ByteArray response in memory and trying to
convert that into a String. We use the Jackson library for streaming.

In the future, we should look into just creating the objects directly
from the stream instead of creating JSON strings.