PatientDao: do not use `@Insert` to do an update
The code uses the `@Insert` functions to do updates for existing
patients.

Using `@Insert` to do an update messes with the foreign keys. Every time
an `@Insert` function is used to update a patient, readings with foreign
keys for that patient all get removed due to the ON DELETE CASCADE
clause. It seems that SQLite's REPLACE just deletes the row and adds it
back? We need a dedicated function that will handle both updates and
inserts when needed.