Patient: change drug and medical history to String
We're given a paragraph field for these inputs, so it doesn't make sense
to assume that they are lists. Furthermore, the backend sends and
accepts Strings for drug and medical history, so there's no point in
doing things differently.

This requires a database migration, since we're changing the column name
and also changing the String format stored in the database.