stop using MainScope() where applicable
MainScope() is not appropriate, because it's creating a new
CoroutineScope that doesn't get cancelled when the Activities are
killed. These coroutines need to be bound to the lifecycle of the
Activity to prevent leaks.

One caveat is the SyncActivity. It should be moved to a service call
independent of the Activity, because it will act strangely if we just
only let users sync if they have SyncActivity open as the active screen.