replace PdfView with TouchImageView for posters
The previous com.github.barteksc:android-pdf-viewer dependency was too
big and took up a lot of space on the APK. For example, on the previous
commit with the com.github.barteksc:android-pdf-viewer dependency, a
debug APK had an app size of ~60.66 MB. With this patch applied, the app
size is now down to ~42.94 MB.

There's no minimal PDF library out there that supports our very-low
minSdkVersion of 17, so using a TouchImageView with pinch-and-zoom
support is a good way to get around that. TouchImageView has
pinch-and-zoom support and a small footprint on the app size.