RestApi: use Jackson for all Http calls
Instead of receiving a response, reading the stream to a ByteArray,
converting that ByteArray into a String, using that String to construct
a JSONObject / JSONArray, and then using that JSONObject to construct
a model class, we use Jackson to read directly from the InputStreams
into the model classes.

Letting Jackson read directly from the InputStream is optimal: "Jackson
is very good at finding the most efficient (sometimes zero-copy) way to
consume/produce JSON encoded data -- allow it do its magic" [1].

For outputs that aren't very big such as a single patient, we just
create a ByteArray for that model object, and from that we use a
fixed-length streaming strategy for uploading. It's a tradeoff, and
we're still experimenting with whether using Jackson to write directly
more efficient than having to deal with a Json object creating a
intermediate String and then converting that into a ByteArray.

[1] https://github.com/FasterXML/jackson-docs/wiki/Presentation:-Jackson-Performance

* Http: support custom OutputStream writing
* SyncUpdate: use Jackson for reading JSON response