implement a test / prototype using Argon2 and AES-GCM
Argon2 is used to derive a key which is then used to encrypt a master
key. This encrypted master key can be encrypted by other keys in the
future (like a key generated from the Android Keystore).

We use the Tink library to work with Cipher objects in javax, because it
is easy to mess up cryptography and it's better to just trust
cryptographers and security engineers to do it for us. In our case, Tink
works with the javax.crypto library for us to handle encryption and
decryption.