extract EncryptedSharedPreferences from androidx.security
The default implementation was too tightly coupled with the MasterKey,
which stores keys in the Android Keystore and falls back to storing keys
in plaintext. We want to see if we can support password-derived keysets,
and it's very much possible to do so since the
EncryptedSharedPreferences implementation only relies on the primitives
to handle to encryption.

The complexity we'd have to manage is the exact arguments that are
required to pass in to the function.