redesign SyncActivity and sync in background
Sync things properly in the background with awareness of internet state.
If the app is killed while syncing, WorkManager will automatically retry
the syncing. There is less informational overload for the user (the
progress shown to the user is only relevant for the current syncing step
and a short summary of what the SyncActivity is for is given).

Other UX improvements include a number of how many patients / readings
are needed to be uploaded and the number of patients and readings
downloaded in the last sync.

Closes MOB-136, closes MOB-132, closes MOB-226