add minimal support for gestational age in days
This fixes syncing with the server.

The current implementation of gestational age days is a minimal working
example. We need to adhere to the proposed design for gestational age
input in the UI, i.e. have a "Weeks and days" option for unit with two
text input boxes: one for weeks and one for days. See tickets MOB-76 and
see CRDL-134 for mockups.