rework NetworkWatcher into LiveData
The previous version using a NetworkWatcher was both Lifecycle-unaware
and didn't catch the case where if the callback is registered while
there is no internet, the LiveData value is still null. To address the
Lifecycle issue, we convert the NetworkWatcher into LiveData, which lets
us use the benefits of the Lifecycle-awareness of LiveData to unregister
our network callbacks when no one is looking for changes.

For checking the initial state of the internet and also to improve the
way we check the connectivity, we now check whether the user has
internet by using deprecated ConnectivityManager methods. Note that the
code inside of androidx.work that handles the network constraints also
uses these deprecated methods. If these deprecated methods finally get
removed, it might be worth to see that they end up doing. Just like
androidx.work, We fall back to a BroadcastReceiver when the
defaultNetworkCallback is not available.

https://issuetracker.google.com/issues/145163549 has a comment from a
Google developer about the usage of these callbacks (although it's
talking about someone else's code):

"Modern Android devices can connect to multiple networks at the same
 time. The NetworkCallback interface will tell you about all the
 networks that match the NetworkRequest you passed, and will pass the
 relevant Network as an argument. Keep in mind that there is no limit to
 how many networks a device can be connected : phones often are
 connected to both cellular and mobile at the same time, and some
 devices can connect to two WiFi networks at the same time for example.

"In your code, you create a default NetworkRequest which will match all
 networks (except the restricted ones and VPNs). So this callback will
 be called for almost all networks, including cellular and WiFi as well
 as Ethernet or Bluetooth networks if present.  When the system calls
 your onAvailable(Network), it will pass the network that is available.

"When the system calls your onLost(Network), it will pass the network
 that was lost. So onLost() doesn't necessarily mean complete loss of
 connectivity, it means one network was lost, and the argument tells you
 which.  However your code ignores the Network argument. If you print it
 in the log, you'll see that the onLost was called for a different
 network. In practice, that means your code assumes loss of connectivity
 when any network at all is lost, even if some other good network is
 still connected.

"In this case, when the device connects to WiFi, the Mobile network may
 stay connected or may disconnect if it's not needed. I suspect the
 onLost you are seeing when you connect to WiFi is the mobile network
 disconnecting as a result of being now unneeded, because the WiFi
 network is better.

"When a network connects, your callback will get a call to onAvailable.
 To figure out what this network is capable of, wait for
 onCapabilitiesChanged, which is guaranteed to immediately come after
 onAvailable (as you already noticed). You can tell whether this is a
 mobile or a WiFi network by querying the NetworkCapabilities with
 hasTransport() and the appropriate transport constant
 (TRANSPORT_CELLULAR or TRANSPORT_WIFI for example).

"Finally, you may want to take a look at
 ConnectivityManager#registerDefaultNetworkCallback. While
 registerNetworkCallback will send you information about all networks
 that match the request you pass, registerDefaultNetworkCallback will
 only tell you about the current default Internet network of the device.
 I suspect this is closer to what you seem to be interested in. Read the
 public documentation for details on how this works."