don't allow network calls in the UI if no internet
https://developer.android.com/topic/performance/power/network/
action-user-traffic#check-or-listen

"Searching for a cell signal is one of the most power-draining
operations on a mobile device. Your app should always check for
connectivity before sending a user-initiated network request. If you use
a scheduling service, Schedulers do this automatically for you.

"If only certain buttons in your activity depend on a network
connection, use Connectivity Manager to check for a network connection
before sending the network request, as instructed in Monitor for Changes
in Connectivity. If there's no network, the app can save battery by not
forcing the mobile radio to search."