CI: add scheduled Firebase Test Lab job for instrumented tests
This is meant to be used with GitLab CI schedules. It will run the
instrumented unit tests using Firebase Test Lab. This requires a GCP
service account with its key which should be provided as variables in
the CI schedule settings.

Instrumented tests in the androidTest directory include database
migration tests, database DAO tests, UI tests, etc. The database
migration tests are especially important, so we want to make sure that
we run them. We only plan to run them on a schedule, since Firebase Test
Lab has quotas for devices on the free plan.

The free plan has a daily quota of 5 test runs with physical devices,
and 10 test runs with virtual devices. See
https://firebase.google.com/docs/test-lab/usage-quotas-pricing.
The quotas reset at midnight PST.