test/loginmanagertests: tweakted test behavior to expect sync timestamp updates
The old test asserts dont make much sense. Timestamps are updated when a download is successful.
Parsing always happen AFTER download, so by the time parsing fails, timestamps are already updated.

The last assert is especially confusing as an earlier assert ensures the network result is successful, so in thoery sync timestamps are always updated by this point