layout/assessment_card: Removed Assessed By from the assessment card
After some discussion, the decision has been to to remove references to healthcare workers

Currently, Android only stores the healthcare worker ID, but not the actual user objects.
This means the system cannot pull out information such as the worker name or their occupation.
One way to approach this would be to allow Android to sync these information from the platform. However, we want to store only the minimum data on Android.
It is better to simply hide the assessed by field than to introduce a new model and a sync endpoint