view: added need activity PatientReferral and its associated view model
After some internal debate the agreed way forward is to leave logic in the referral dialog alone, as it is too deeply tangled with reading
By adding a new activity, this opens up the freedom for expansion. We can expand the referral creation UI and attach more viewmodel endpoints
It also makes passing data around much simpler, as intent extra are much easier to work with.